move fetch to own module
